# more hyper-rust optimizations