--- a/rust-hyper/execution-engine/Cargo.toml
+++ b/rust-hyper/execution-engine/Cargo.toml
@@ -13,5 +13,6 @@ itertools = "0.9.0"
 ramp = "0.5.9"
 macros = { version = "*", path = "../macros"}
 serde_json = "1.0"
-chttp = "0.5.5"
+reqwest = {version = "0.10", features = ["blocking"]}
 serde = "1.0"
+lazy_static = "1.4"