* Use a pure-rust http client (AFAIK chttp uses libcurl, and reqwest's author is hyper's author)
* instead of panicking (and killing a thread), return an error when dividing/modulus by zero
* Save more allocations in program construction by using static arrays instead of vectors
* `spawn_blocking` should be used when spawning a blocking task (such as `eval`)