@pbiggar 2020-09-30
Thanks!

> Save more allocations in program construction by using static arrays instead of vectors

Can you confirm this makes it faster (or at least, no slower) for you?

> spawn_blocking should be used when spawning a blocking task (such as eval)

I don't understand this change. Could you explain it for the next person to work on this (as a comment if appropriate, else here is fine)?

Thanks!