@naim94a 2020-09-30
>> Save more allocations in program construction by using static arrays instead of vectors
>
> Can you confirm this makes it faster (or at least, no slower) for you?

Yes, I used [`criterion`](https://github.com/bheisler/criterion.rs) to benchmark the `eval` function.

Current main branch:
```
fizzbuzz                time:   [663.04 us 679.78 us 697.84 us]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```
```
fizzbuzz                time:   [637.85 us 653.05 us 668.90 us]
                        change: [-9.0754% -4.4719% +0.5992%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
```

This PR:
```
fizzbuzz                time:   [473.37 us 487.92 us 503.96 us]
                        change: [-28.994% -25.821% -22.838%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

If you'd like I can include the benchmark's code in a separate PR.

>> spawn_blocking should be used when spawning a blocking task (such as eval)
>
> I don't understand this change. Could you explain it for the next person to work on this (as a comment if appropriate, else here is fine)?

Tasks spawned on the `tokio` executor must not block. `rust-hyper` will block when making HTTP requests, which may block the executor and cause a delay. 
From [`spawn_blocking`](https://docs.rs/tokio/0.2.22/tokio/task/fn.spawn_blocking.html)'s documentation:
> Runs the provided closure on a thread where blocking is acceptable.